01: 5: fixed types for target, started well on Mac

+++ 01_unity_proto/proto_5/Target.cs
@@ -56,7 +56,7 @@ public class Target : MonoBehaviour
 
         if (Target.prevXPos != null)
         {
-            while (System.Math.Abs(xPos - Target.prevXPos) < 1f)
+            while (System.Math.Abs(xPos - (float)Target.prevXPos) < 1f)
             {
                 xPos = Random.Range(-xRange, xRange);
             }
